# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Book:
    properties:
      book_id:
        type: string
      cover_url:
        description: null when the book has no cover
        type: string
      description:
        type: string
      is_public:
        type: boolean
      title:
        type: string
    type: object
  models.BookIDResponse:
    properties:
      book_id:
        type: string
    type: object
...
  models.CreateBookBody:
    properties:
      author_id:
        type: string
      description:
        type: string
      language_id:
        type: string
      published_at:
        type: string
      title:
        type: string
    required:
    - author_id
    - description
    - language_id
    - published_at
    - title
    type: object
  models.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  models.UpdateBookBody:
    properties:
      author_id:
        type: string
      description:
        minLength: 1
        type: string
      language_id:
        type: string
      published_at:
        type: string
      title:
        minLength: 1
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Readness API
  version: "1.0"
paths:
...
  /books:
    get:
      description: Публичные книги с фильтрами и поиском по названию, по 20 на страницу
      parameters:
      - description: Поиск по названию
        in: query
        name: q
        type: string
      - description: ID жанра
        in: query
        name: topic_id
        type: string
      - description: ID языка
        in: query
        name: language_id
        type: string
      - description: ID автора
        in: query
        name: author_id
        type: string
      - description: Страница, с 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Список книг
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Создаёт непубличную книгу; публикация — через заявку на модерацию
      parameters:
      - description: Данные книги
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создание книги
      tags:
      - Books
  /books/{id}:
    delete:
      description: Владелец или модератор
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Удаление книги
      tags:
      - Books
    get:
      description: Публичная книга видна всем, непубличная — только владельцу и модераторам
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Книга
      tags:
      - Books
    patch:
      consumes:
      - application/json
      description: Владелец может править пока книга не опубликована, модератор —
        всегда. Меняются только переданные поля
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      - description: Изменяемые поля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Редактирование книги
      tags:
      - Books
  /books/{id}/cover:
    get:
      description: Видимость как у самой книги
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Book or cover not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обложка книги
      tags:
      - Books
    put:
      consumes:
      - multipart/form-data
      description: 'multipart/form-data, поле cover: JPEG, PNG или WebP до 2 МБ. Права
        как у редактирования книги. Заменяет прежнюю обложку'
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      - description: Файл обложки
        in: formData
        name: cover
        required: true
        type: file
      responses:
        "204":
          description: Обложка сохранена
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Загрузка обложки
      tags:
      - Books
...
  /users/me/books:
    get:
      description: Книги из библиотеки пользователя, опционально по статусу
      parameters:
      - description: Статус
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Моя библиотека
      tags:
      - Books
swagger: "2.0"